lib/backend/app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Antoine Vandermeersch <avdmeers@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aws-lambda-powertools = {extras = ["tracer"], version = "^2.16.2"}
auth0-python = "^4.2.0"
langchain = "^0.0.206"
openai = "^0.27.8"
pydantic = "^1.10.9"


[tool.poetry.group.dev.dependencies]
ipython = "^8.14.0"
boto3 = "^1.26.156"
pytest = "^7.3.2"
pytest-cov = "^4.1.0"
coverage = {extras = ["toml"], version = "^7.2.7"}
pytest-env = "^0.8.2"


[tool.poetry.group.lint.dependencies]
flake8 = "^6.0.0"
flake8-black = "^0.3.6"
flake8-import-order = "^0.18.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.coverage.paths]
source = ["src", "*/site-packages"]

[tool.coverage.run]
branch = true
source = ["app"]

[tool.coverage.report]
show_missing = true

[tool.pytest.ini_options]
env = [
  "GAME_TABLE=DummyGameTable",
  "QUESTION_TABLE=DummyQuestionTable",
]
